fix: Move force_path_style to use_path_style
As per the documentation linked below the force_path_style option is now
deprecated and should be changed to use_path_style.

https://developer.hashicorp.com/terraform/language/backend/s3#force_path_style

It may be better to add both to allow people to migrate from one to the
other but given this is only in one place in the docs I expect there
isn't a lot of use and highlighting as a breaking change will make
future migration easier and stop being using the now deprecated option.